pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "azure-upload-benchmark"
version = "0.1.0"
description = "Benchmark Azure Blob Storage upload performance with varying chunk sizes and concurrency levels."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Pranaya Shrestha", email = "pranaya.shrestha2@gmail.com"}]
keywords = ["azure", "blob", "storage", "benchmark", "upload", "performance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Benchmark",
]
dependencies = [
    "azure-storage-blob>=12.20.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
plots = [
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
]
dev = [
    "azure-upload-benchmark[plots]",
    "pytest>=9.0",
    "pytest-cov>=7.1",
]

[project.scripts]
azure-upload-benchmark = "azure_upload_benchmark.cli:main"

[project.urls]
Homepage = "https://github.com/pshresth/azure-upload-benchmark"
Repository = "https://github.com/pshresth/azure-upload-benchmark"
Issues = "https://github.com/pshresth/azure-upload-benchmark/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_upload_benchmark"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
